+++ rust-toolchain.toml
@@ -1,0 +1,3 @@
+ [toolchain]
+  channel = "nightly-2023-05-27"
+  components = ["rust-src", "rustc-dev", "llvm-tools"]
