+++ src/main.rs
@@ -1,14 +1,8 @@
-use math::Vec3;
-use voxel::brick_map::BrickMap;
-
 mod math;
-mod colors;
-mod texture;
 mod camera;
 mod application;
 mod rendering;
 mod voxel;
-mod debug_utils;
 mod utils;
 pub mod gpu_utils;
