+++ src/shaders/test_shader/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "test_shader"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["dylib"]
+
+[dependencies]
+spirv-std = { version = "0.9" }
